toJsonString()への切り出し、fetchWeather()の返り値なしへの変更、getWeatherIcon()への切り出し
