+++ Training/Training/ViewController.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 import YumemiWeather
+import Foundation
 
 //APIに渡すJson文字列の器
 struct ServeInfo: Encodable {
@@ -22,47 +23,73 @@ struct ReceiveInfo: Decodable {
     var date: String
 }
 
+//プロトコル
+protocol forecastDelegate: AnyObject {
+    func toJsonString() -> String?
+    func fetchWeather()
+    func getWeatherIcon() -> UIImage?
+}
+
 //AlertController表示に使用する変数
 var errorMessage: String?
 //取得した情報(オブジェクト形式)を格納する変数
 var receiveInfo: ReceiveInfo? = nil
 
-//プロトコル
-protocol forecastDelegate: AnyObject {
-    func fetchWeather() -> UIImage?
-}
-
 //処理内容を記した、処理を任されるクラスその1
 class YumemiForecast: forecastDelegate {
     
-    func fetchWeather() -> UIImage? {
-        var weather: String?
+    var jsonData: Data?
+    
+    //オブジェクトからJson形式へ変換（エンコード）
+    func toJsonString() -> String? {
+        let serveInfo = ServeInfo(area: "tokyo", date: "2020-04-01T12:00:00+09:00")
+        let encoder = JSONEncoder()
         do {
-            errorMessage = nil
-            //オブジェクトからJson形式へ変換（エンコード）
-            let serveInfo = ServeInfo(area: "tokyo", date: "2020-04-01T12:00:00+09:00")
-            let encoder = JSONEncoder()
-            encoder.outputFormatting = .prettyPrinted
-            var jsonData = try encoder.encode(serveInfo)
-            let jsonString = String(data: jsonData, encoding: .utf8)!
-            //Json文字列を引数にAPI呼び出し、天気取得
-            try weather = YumemiWeather.fetchWeather(jsonString)
-            //受け取ったJson文字列をオブジェクトに変換（デコード）
-            if let weatherData = weather {
-                jsonData = weatherData.data(using: .utf8)!
-            }
-            receiveInfo = try JSONDecoder().decode(ReceiveInfo.self, from: jsonData)
-        } catch (YumemiWeatherError.invalidParameterError) {
-            errorMessage = "invalidParameterErrorが発生しました"
-            return nil
-        } catch (YumemiWeatherError.unknownError) {
-            errorMessage = "unknownErrorが発生しました"
-            return nil
+            jsonData = try encoder.encode(serveInfo)
         } catch {
             errorMessage = "予期せぬエラーが発生しました"
             print(error)
             return nil
         }
+        if let jsonData = jsonData {
+            let jsonString = String(data: jsonData, encoding: .utf8)!
+            return jsonString
+        } else {
+            return nil
+        }
+    }
+    
+    //API取得、デコード
+    func fetchWeather() {
+        var weather: String?
+        do {
+            errorMessage = nil
+            receiveInfo = nil
+            //オブジェクトからJson形式へ変換（エンコード）
+            let jsonString = toJsonString()
+            if let jsonString = jsonString {
+                //Json文字列を引数にAPI呼び出し、天気取得
+                try weather = YumemiWeather.fetchWeather(jsonString)
+            }
+            //受け取ったJson文字列をオブジェクトに変換（デコード）
+            if let weatherData = weather {
+                jsonData = weatherData.data(using: .utf8)!
+            }
+            if let jsonData = jsonData {
+                receiveInfo = try JSONDecoder().decode(ReceiveInfo.self, from: jsonData)
+            }
+        } catch (YumemiWeatherError.invalidParameterError) {
+            errorMessage = "invalidParameterErrorが発生しました"
+        } catch (YumemiWeatherError.unknownError) {
+            errorMessage = "unknownErrorが発生しました"
+        } catch {
+            errorMessage = "予期せぬエラーが発生しました"
+            print(error)
+        }
+    }
+    
+    //画像を取得
+    func getWeatherIcon() -> UIImage? {
         if let weatherResult = receiveInfo {
             switch weatherResult.weather_condition {
             case "sunny":
@@ -86,34 +113,12 @@ class Forecast {
     //weak必須
     weak var delegate: forecastDelegate? = nil
     
-    func doFetchWeather() -> UIImage? {
+    func doFetchWeather()  {
         if let dg = self.delegate {
             return dg.fetchWeather()
         } else {
-            return nil
+            print("delegate実行不可")
         }
-    }
-    
-}
-
-//UIArertControllerを生成するクラス
-class CreateAlertController {
-    
-    func create (_ message: String) -> UIAlertController {
-        // UIAlertControllerの生成
-        let alert = UIAlertController(title: message, message: "再試行してください", preferredStyle: .alert)
-        // アクションの生成
-        let yesAction = UIAlertAction(title: "了解", style: .default) { action in
-            //追加処理なし
-        }
-        let noAction = UIAlertAction(title: "納得できません", style: .destructive) { action in
-            //追加処理なし
-        }
-        // アクションの追加
-        alert.addAction(yesAction)
-        alert.addAction(noAction)
-        
-        return alert
     }
     
 }
@@ -130,12 +135,13 @@ class ViewController: UIViewController {
     @IBOutlet var minTemperature: UILabel!
     
     @IBAction func fetchWeather(_ sender: Any) {
+        var weatherIcon: UIImage? = nil
         //処理を任せるクラスのインスタンス生成
         let forecast = Forecast()
         //今回処理を任されるクラスのインスタンス生成 と紐付け
         let yumemiForecast = YumemiForecast()
         forecast.delegate = yumemiForecast
-        let weatherIcon: UIImage? = forecast.doFetchWeather()
+        forecast.doFetchWeather()
         
         //exceptionルートを通っていたらUIArertControllerでエラー表示
         if let message = errorMessage  {
@@ -145,12 +151,12 @@ class ViewController: UIViewController {
             // UIAlertControllerの表示
             present(alertController, animated: true, completion: nil)
         } else {
-
-            //exceptionのルートを通っていなかったら天気画像を表示
+            //exceptionのルートを通っていなかったら天気画像等を表示
+            if let receiveInfoExist = receiveInfo {
+            weatherIcon = yumemiForecast.getWeatherIcon()
             if let icon = weatherIcon {
                 weather.image = icon
             }
-            if let receiveInfoExist = receiveInfo {
                 maxTemperature.text = "\(receiveInfoExist.max_temperature)"
                 minTemperature.text = "\(receiveInfoExist.min_temperature)"
             }
